found how to not have overflow

+++ tarea1/include/rtree/HilbertCurve.hpp
@@ -7,7 +7,7 @@
 
 using namespace std;
 
-int hilbertValue(Point point, int n);
+long long hilbertValue(Point point, long long n);
 Node* hilbertCurve(int M, vector<Rectangle> rectangles);
 
 #endif // NEAREST_X_HPP
